requirements.txt
streamlit 
scikit-learn 
google-generativeai 
numpy
dotenv
pdfminer.six